+++ src/main/java/io/github/yokigroup/event/submodule/abs/PlayerCharacterSubmoduleAbs.java
@@ -45,7 +45,7 @@ public abstract class PlayerCharacterSubmoduleAbs extends Submodule {
     public abstract void movePlayerBy(Vector2 delta);
 
     @Override
-    protected void updateCode(long delta) {
+    protected void updateCode(double delta) {
         // TODO collision check
     }
 }
